feat: complete modernization and architectural fixes (Issue #7562, #7801, #6436, #6822, #7542)
